Split up main.go
